Getting data
Because node.js is non-blocking, when getting a response, we get
multiple chunks of data, as they are streamed in through the network.
This lets us do other things whilst getting data from the network.

Only when the response calls an end event, do we know that there is no
more data to be collected, and we can now print out the entire body in
one.