Handling 404 errors.
If the page isn't found, we'll handle it. Also created printError()
function to deal with multiple errors, to stay DRY :)

Note that we're getting the status code from the http.STATUS_CODES
array, NOT https (Not an existant array), therefore had to include http.
This just gives us a friendly descriptor for the error. e.g: 404 === Not
found.